# [PERF] stock: Batching get_rules_from_location