Currrently `product.product._get_rules_from_location` can become a performance bottleneck when it's called multiple times. Because the method finds a candidate stock.rule then call itself recursively with stock.rule.location_src_id, it's difficult to properly batch.

In this commit we introduce two methods,
`product.product._get_rules_for_combinations` and
`procurement.group._get_rules_for_combinations`.

The second one is a batched version of `get_rule`. The idea is that given a list of (products, locations, warehouses), the method will call `search_rules_for_warehouses` only once and then distribute the fetched stock.rule to the correct `(product, location, warehouse)` triplet by building a dictionary.

This is then used by `product.product.get_rules_for_combinations` which takes a list of `(products, locations)` pairs and build a recordset of stock.rules for each one.

The way to use this workflow is the following:
 - The calling code creates the list of (product_id, location_id) it needs. E.g. `[(o.product_id, o.location_id) for o in orderpoints)]`
 - A single call to `product.product._get_rules_for_combinations`
 - stock.rules are then extracted by the calling code using the resulting dictionary.

Assuming the worst case, the current code does (n+k) calls to both `product.product._get_rules_from_location` and `procurement.group._get_rule`, with n being
`len([(r.product_id, r.location_id) for r in records])` and k the total number of recursive calls in
`product.product._get_rules_from_location`

After this commit, the worst case will be h calls to both methods, with h being the height of
`product.product._get_rules_for_combinations` recursive call tree.

#### speedup

Current iterative version product._get_rules_from_location 
| Total loops count |  Exec Time |
|:-------------------:|:------------:|
|        10         |     50ms   |
|        50         |     62ms   |
|       500         |     567ms  |
|     7 441         |     5.6s   |

Benchmark of product._get_rules_for_combinations
| combinations Nb | Exec Time |
|:-------------------:|:-----------:|
|       10        |   24 ms   |
|       50        |   23ms    |
|      500        |   93ms    |
|    7 441        |   1.23s   |

Average speedup calling both method from a shell: 2.4

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr